+++ modules/openapi-generator/src/test/java/org/openapitools/codegen/cpptiny/CppTinyClientModelTest.java
@@ -1,0 +1,17 @@
+package org.openapitools.codegen.cpptiny;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.media.IntegerSchema;
+import io.swagger.v3.oas.models.media.ObjectSchema;
+import io.swagger.v3.oas.models.media.Schema;
+import org.openapitools.codegen.*;
+import org.openapitools.codegen.languages.CppTinyClientCodegen;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+
+
+public class CppTinyClientModelTest extends CppTinyBaseTest {
+
+
+}
